+++ Source/ProjectRimFactory/Drones/Pawn_Drone.cs
@@ -11,56 +11,6 @@ using UnityEngine;
 namespace ProjectRimFactory.Drones
 {
 
-    public class DroneArea : Area
-    {
-        private string labelInt;
-        public DroneArea(AreaManager areaManager, string label = null) : base(areaManager)
-        {
-            base.areaManager = areaManager;
-            if (!label.NullOrEmpty())
-            {
-                labelInt = label;
-            }
-            else
-            {
-                int num = 1;
-                while (true)
-                {
-                    labelInt = "AreaDefaultLabel".Translate(num);
-                    if (areaManager.GetLabeled(labelInt) == null)
-                    {
-                        break;
-                    }
-                    num++;
-                }
-            }
-            colorInt = new Color(Rand.Value, Rand.Value, Rand.Value);
-            colorInt = Color.Lerp(colorInt, Color.gray, 0.5f);
-        }
-
-
-        private Color colorInt = Color.red;
-
-        private string LabelText = "DroneZone";
-
-        public override string Label => LabelText;
-
-        public override Color Color => colorInt;
-
-        public override int ListPriority => throw new NotImplementedException();
-
-        public override string GetUniqueLoadID()
-        {
-            throw new NotImplementedException();
-        }
-
-    }
-
-    public class DroneDefModExtension : DefModExtension
-    {
-        public int SquareJobRadius = 0; //0 Means infinite
-    }
-
     public class Pawn_Drone : Pawn
     {
         public Building_DroneStation station;
@@ -108,25 +58,25 @@ namespace ProjectRimFactory.Drones
             Name = new NameSingle("PRFDroneName".Translate());
 
 
-
             //If range is set (bigger then 0) then do handel the range
-            if (this.station.def.GetModExtension<DroneDefModExtension>().SquareJobRadius > 0) { 
-                //Handel Allowed area (not to be confuced with Area_Allowed)
-                Area droneArea;
-                droneArea = new DroneArea(this.Map.areaManager);
-                //Need to set the Area to a size
-
-                foreach (IntVec3 cell in Rangecells)
+                if (this.station.def.GetModExtension< DroneDefModExtension>().SquareJobRadius > 0)
                 {
-                    droneArea[cell] = true;
+                    //Handel Allowed area (not to be confuced with Area_Allowed)
+                    Area droneArea;
+                    droneArea = new DroneArea(this.Map.areaManager);
+                    //Need to set the Area to a size
+
+
+                    foreach (IntVec3 cell in Rangecells)
+                    {
+                        droneArea[cell] = true;
+                    }
+                    //Not shure if i need that but just to be shure
+                    droneArea[Position] = true;
+
+                    playerSettings.AreaRestriction = droneArea;
                 }
-                //Not shure if i need that but just to be shure
-                droneArea[Position] = true;
-
-                playerSettings.AreaRestriction = droneArea;
-            }
             
-
 
         }
